fix build error in build.gradle
Potential workaround for the error with jna-platform.jar.

The issue is that the mariadb-java-client POM has two dependencies.
One is for net.java.dev.jna:jna with the classifier and one is for
without. And since 4.0, net.java.dev.jna:jna publishes the
'platform' artifact under a different name
(net.java.dev.jna:jna-platform).
(Source: "https://discuss.gradle.org/t/build-failure-when-4-1-0-of-
net-java-dev-jna-jna-is-selected-in-place-of-3-3-0-due-to-change-
in-artifacts-with-classifiers/9377"